wip: replace old text input element
