--- webui/src/Components/WindowedSelect/README.md
@@ -1,3 +1,0 @@
-Adapted from https://github.com/jacobworrel/react-windowed-select
-
-License: MIT
